% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to a database instance}
\usage{
connect(dsn, uid = NULL, pwd = NULL)
}
\arguments{
\item{dsn}{The dsn name, as it appears in your \code{tnsnames.ora}}

\item{uid}{User ID - if NULL, it will be looked up in the getcdw encrypted credential databse}

\item{pwd}{Password - if NULL, it will be looked up in the getcdw encrypted credential databse}
}
\description{
When running \code{\link{get_cdw}} and related functions, the connection to
the database is opened silently in the background, but this function might
be useful if you need to create a connection without immediately querying
the database. For instance, \url{http://rmarkdown.rstudio.com/authoring_knitr_engines.html#sql}.
}

