% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-codes.R
\name{find_codes}
\alias{find_codes}
\title{Find codes in the tms table}
\usage{
find_codes(search_term = NULL, table_name = NULL, ...)
}
\arguments{
\item{search_term}{A search term (can be an Oracle SQL regular expression)}

\item{table_name}{TMS table name (or regular expression)}

\item{...}{Any parameters passed on to \code{get_cdw}}
}
\description{
Use regular expression matching to look for terms in different
code tables. For instance, \code{find_codes("astronomy")} will return major,
minor, department, etc. codes.
}
\details{
Either \code{search_term} or \code{table_name} can be \code{NULL}
}

