% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-columns.R
\name{find_columns}
\alias{find_columns}
\title{Search a database schema for column names}
\usage{
find_columns(search_term = NULL, table_name = NULL, schema = "CDW")
}
\arguments{
\item{search_term}{Pattern of the column(s) you are looking for, e.g. 'class_campaign' or 'household'}

\item{table_name}{Name (or partial name) of the table (optional)}

\item{schema}{In case you're searching through something other than the CDW, e.g. 'advance'}
}
\description{
Search a database schema for column names
}

