% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-tables.R
\name{find_tables}
\alias{find_tables}
\title{Search a database schema for tables}
\usage{
find_tables(search_term = NULL, schema = "CDW", ...)
}
\arguments{
\item{search_term}{Pattern of the table(s) you are looking for, e.g. 'committee'}

\item{schema}{In case you're searching through something other than the CDW, e.g. 'advance'}

\item{...}{Passed on to \code{get_cdw}}
}
\description{
Search a database schema for tables
}
