% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-cdw.R
\name{get_cdw}
\alias{get_cdw}
\title{Get data from the CDW}
\usage{
get_cdw(query, dsn = "CDW2", uid, pwd, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{query}{A SQL query, as a string of plain-text.}

\item{dsn}{The name of the connection, as it appears in your \code{tnsnames}}

\item{uid}{Your username}

\item{pwd}{Your password}

\item{stringsAsFactors}{TRUE/FALSE. Whether strings should be returned as factors. FALSE by default}

\item{...}{Other arguments passed on to \code{sqlQuery} in package \code{RODBC}}
}
\description{
Get data from the CDW
}
\details{
Returns a data.frame (unless)
}

