% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize-template.R
\name{parameterize_template}
\alias{parameterize_template}
\title{Create a template with parameters}
\usage{
parameterize_template(tmpl)
}
\arguments{
\item{tmpl}{a template, where parameters are ##highlighted## with double-octothorpes}
}
\description{
Create a template with parameters
}
\examples{
greeting_template <- "my name is ##name##, and I am ##age## years old"
greeter <- parameterize_template(greeting_template)
greeter(name = "tarak", age = 36)
}
