% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-cdw.R
\name{preview_cdw}
\alias{preview_cdw}
\title{Run a preview query}
\usage{
preview_cdw(query, dsn = "CDW2", uid = NULL, pwd = NULL,
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{query}{A SQL query, as a string of plain-text.}

\item{dsn}{The name of the connection, as it appears in your \code{tnsnames}}

\item{uid}{Your username (see details)}

\item{pwd}{Your password (see details)}

\item{stringsAsFactors}{TRUE/FALSE. Whether strings should be returned as factors. FALSE by default}

\item{...}{Other arguments passed on to \code{sqlQuery} in package \code{RODBC}}
}
\description{
Run a query but only return the first 10 rows of results. Useful for testing
queries that might run slowly.
}

