% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-preview-query.R
\name{runPreviewQuery}
\alias{runPreviewQuery}
\title{Run a preview query}
\usage{
runPreviewQuery()
}
\description{
In RStudio, use this as an add-in to interactively run a preview query
against the data warehouse. Returns the first 10 rows. Bind a keyboard
shortcut through the RStudio "Addins | Browse Addins | Keyboard shortcuts"
menu. RStudio also has syntax highlighting/indenting for SQL files (adjust
the file type in the bottom right corner of the text editor window).
}

