% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cdw.R
\name{get_cdw}
\alias{get_cdw}
\title{Get data from the CDW}
\usage{
get_cdw(query, dsn = "CDW2", uid = NULL, pwd = NULL,
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{query}{A SQL query. This can be either the text of the query, the name
of a file that contains the query, or a valid connection that contains the
text of the query.}

\item{dsn}{The name of the connection, as it appears in your \code{tnsnames}}

\item{uid}{Your username (see details)}

\item{pwd}{Your password (see details)}

\item{stringsAsFactors}{TRUE/FALSE. Whether strings should be returned as factors. FALSE by default}

\item{...}{Other arguments passed on to \code{sqlQuery} in package \code{RODBC}}
}
\description{
Get data from the CDW
}
\details{
Returns a data.frame if the query is successful, otherwise an error.
If you don't enter a username/password, \code{get_cdw} will check the global
environment for these values. The first time you run a query against a given
\code{dsn}, you will be prompted for your credentials. These will be saved in
an encrypted database in your home directory, so that you won't be prompted
for them again in the future. If you need to reset your credentials, use the
function \link{\code{reset_credentials}}.
}

