% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cdw.R
\name{preview_cdw}
\alias{preview_cdw}
\title{Run a preview query}
\usage{
preview_cdw(query, n = 10, dsn = config("default_dsn"), uid = NULL,
  pwd = NULL, ...)
}
\arguments{
\item{query}{A SQL query. This can be either the text of the query, the name
of a file that contains the query, or a valid connection that contains the
text of the query.}

\item{n}{The maximum number of rows to return}

\item{dsn}{The name of the connection, as it appears in your \code{tnsnames}}

\item{uid}{Your username (see details)}

\item{pwd}{Your password (see details)}

\item{...}{Other arguments passed on to \code{dbSendQuery} in package \code{ROracle}}
}
\description{
Run a query but only return the first n rows of results. Useful for testing
queries that might run slowly.
}

