% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{reset_credentials}
\alias{reset_credentials}
\title{Set/reset credentials}
\usage{
reset_credentials(dsn)
}
\arguments{
\item{dsn}{The name of the connection, as it appears in your tnsnames}
}
\description{
If run in an interactive session, you will be prompted for your UID (user ID)
and PWD (password). Your entries will be stored in an environment variable,
as well as saved in the encrypted database called ".getcdw" in your home
direcotry.
}

